pyproject.toml
[tool.poetry]
name = "musicbot"
version = "0.7.5"
description = "Music swiss army knife"
authors = ["Adrien Pensart <crunchengine@gmail.com>"]
maintainers = ["Adrien Pensart <crunchengine@gmail.com>"]
homepage = "https://github.com/AdrienPensart/musicbot"
license = "MIT"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
musicbot = 'musicbot.main:main'

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
logging-tree = "^1.8"
pyacoustid = "^1.1"
python-vlc = "^3.0"
prompt_toolkit = "^3.0.2"
mutagen = "^1.44.0"
requests = "^2.24.0"
spotipy = "^2.16.0"
youtube_dl = "^2021.3.3"
watchdog = "^2.0.2"
pydub = "^0.25.1"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.2"
attrs = "^21.2.0"
types-requests = "^2.25.0"
colorlog = "^6.4.1"
types-python-slugify = "^5.0.0"
progressbar2 = "^4.0.0"
rich = "^12.0.0"
click-skeleton = "^0.21"
beartype = "^0.10.0"
humanize = "^4.0.0"
python-slugify = "^6.0.1"
deepdiff = "^5.7.0"
edgedb = "^0.23.0"

[tool.isort]
multi_line_output = 3

[tool.poetry.dev-dependencies]
coverage-badge = "^1.0"
restructuredtext_lint = "^1.3.0"
bump2version = "^1.0.0"
pylint = "^2.6.0"
pytest-docker-compose = "^3.1.2"
docker = { version = "^5.0", markers = "sys_platform != 'win32'" }
pytest-cov = "^3.0.0"
pytest-timeout = "^2.0.1"
flake8 = "^4.0.1"
mypy = "^0.941"
pytest-click = "^1.0.2"
isort = "^5.10.1"
bandit = "^1.7.2"
safety = "^1.10.3"
vulture = "^2.3"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--timeout 600 --cov-report term-missing --durations=0 --cov musicbot --docker-compose-no-build -x --use-running-containers "

[tool.pylint.master]
jobs = 4
[tool.pylint.messages_control]
disable = """
disable=too-few-public-methods,logging-format-interpolation,line-too-long,too-many-arguments,protected-access,missing-docstring,invalid-name,too-many-public-methods,too-many-instance-attributes,duplicate-code,too-many-nested-blocks,too-many-branches,too-many-return-statements,too-many-statements,too-many-locals,too-few-public-methods,too-many-ancestors,abstract-method,anomalous-backslash-in-string,import-outside-toplevel,redefined-outer-name,unnecessary-lambda,c-extension-no-member,logging-fstring-interpolation,too-many-boolean-expressions
"""
